@@ -8,40 +8,23 @@
   },
   "private": true,
   "dependencies": {
-    "@craco/craco": "^6.1.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router-dom": "^5.1.7",
-    "axios": "^0.21.2",
-    "intersection-observer": "^0.12.0",
-    "react": "^17.0.2",
-    "react-cool-img": "^1.2.9",
-    "react-dom": "^17.0.2",
-    "react-feather": "^2.0.9",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.1.2",
-    "web-vitals": "^1.0.1"
+    "axios": "^0.27.2",
+    "intersection-observer": "^0.12.2",
+    "react": "^18.2.0",
+    "react-cool-img": "^1.2.12",
+    "react-dom": "^18.2.0",
+    "react-feather": "^2.0.10",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "craco start",
     "build": "craco build",
-    "test": "craco test",
+    "test": "react-scripts test",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
-    "eject": "react-scripts eject",
-    "prepare": "husky install"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "eject": "react-scripts eject"
   },
   "browserslist": {
     "production": [
@@ -56,27 +39,27 @@
     ]
   },
   "devDependencies": {
-    "@commitlint/cli": "^12.1.1",
-    "@commitlint/config-conventional": "^12.1.1",
+    "@commitlint/cli": "^17.0.3",
+    "@commitlint/config-conventional": "^17.0.3",
+    "@craco/craco": "^6.4.4",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.1.8",
     "@types/axios": "^0.14.0",
-    "@typescript-eslint/eslint-plugin": "^4.4.1",
-    "autoprefixer": "^9",
-    "eslint-config-airbnb-typescript": "^12.3.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.20.3",
-    "eslint-plugin-react-hooks": "^4.0.8",
-    "gh-pages": "^3.1.0",
-    "husky": "^6.0.0",
+    "@types/jest": "^28.1.4",
+    "@types/node": "^18.0.1",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.5",
+    "@types/react-router-dom": "^5.3.3",
+    "autoprefixer": "^10.4.7",
+    "gh-pages": "^4.0.0",
     "make-coverage-badge": "^1.2.0",
     "msw": "^0.28.2",
-    "postcss": "7",
-    "prettier": "^2.2.1",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat",
-    "tailwindcss-debug-screens": "^2.0.0"
+    "postcss": "8.4.14",
+    "react-scripts": "5.0.1",
+    "tailwindcss": "3.1.4",
+    "tailwindcss-debug-screens": "^2.2.1",
+    "typescript": "^4.7.4"
   },
   "jest": {
     "coverageReporters": [
