@@ -1,7 +1,7 @@
 import Routes from './routes/Routes';
 
-const App = () => {
+function App() {
   return <Routes />;
-};
+}
 
 export default App;
