@@ -1,15 +1,14 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import getAllCountries200 from 'mocks/responses/getAllCountries200.json';
-import { MemoryRouter, withRouter } from 'react-router';
+import { MemoryRouter } from 'react-router';
 import userEvent from '@testing-library/user-event';
 import CountryDetails from './CountryDetails';
 
 const country = getAllCountries200[0];
 const asyncRender = async () => {
-  const Wrapper = withRouter(CountryDetails);
   render(
     <MemoryRouter initialEntries={[`/${country.alpha3Code}`]}>
-      <Wrapper />
+      <CountryDetails />
     </MemoryRouter>
   );
 
